+++ packages/studio-base/src/panels/ThreeDeeRender/renderables/PointCloudsAndLaserScans.ts
